Rename nuimo module
