+++ nuimo.py
